Moves dtos to domain based structure
