+++ apps/frontend/src/app/components/organisms/view-event-general.component.ts
@@ -1,7 +1,7 @@
 import { JsonPipe, NgIf } from '@angular/common';
 import { ChangeDetectionStrategy, Component, Input } from '@angular/core';
 import { NbCardModule } from '@nebular/theme';
-import { DetailPlanDto } from '../../models/detail-plan.dto';
+import { DetailPlanDto } from '../../models/plan/detail-plan.dto';
 import { EventOverviewComponent } from '../molecules/event-overview.component';
 import { SelectGatheringComponent } from '../molecules/select-gathering.component';
