+++ apps/frontend/src/app/components/pages/login.component.ts
@@ -3,8 +3,8 @@ import { ChangeDetectionStrategy, Component } from '@angular/core';
 import { Router } from '@angular/router';
 import { NbTabsetModule } from '@nebular/theme';
 import { ROUTE_EVENTS } from '../../constants';
-import { LoginUser } from '../../models/login-user.dto';
-import { RegisterUser } from '../../models/register-user.dto';
+import { LoginUser } from '../../models/user/login-user.dto';
+import { RegisterUser } from '../../models/user/register-user.dto';
 import { AuthService } from '../../services/auth.service';
 import { LoginFormComponent } from '../organisms/login-form.component';
 import { RegisterFormComponent } from '../organisms/register-form.component';
