+++ apps/frontend/src/app/components/pages/plan-event.component.ts
@@ -11,8 +11,8 @@ import { NbButtonModule, NbStepperModule } from '@nebular/theme';
 import { Observable, combineLatest, map } from 'rxjs';
 import { ROUTE_VIEW_EVENT } from '../../constants';
 import { MOCK_GAME_DTOS_LARGE } from '../../mocks/game.mock';
-import { CreatePlan } from '../../models/create-plan.dto';
-import { Game } from '../../models/game.dto';
+import { Game } from '../../models/game/game.dto';
+import { CreatePlan } from '../../models/plan/create-plan.dto';
 import { PlanService } from '../../services/plan.service';
 import { get24HourTime } from '../../utils/date.utility';
 import { PlanEventSummaryComponent } from '../molecules/plan-event-summary.component';
