+++ apps/frontend/src/app/models/gathering/create-gathering.dto.ts
@@ -1,5 +1,5 @@
-import { Model } from '../utils/types';
-import { Dto } from './base.dto';
+import { Model } from '../../utils/types';
+import { Dto } from '../base.dto';
 import { GatheringDto } from './gathering.dto';
 
 /**
