+++ apps/frontend/src/app/models/gathering/detail-gathering.dto.ts
@@ -1,5 +1,5 @@
-import { Model } from '../utils/types';
-import { Dto } from './base.dto';
+import { Model } from '../../utils/types';
+import { Dto } from '../base.dto';
 import { Gathering } from './gathering.dto';
 
 /**
