+++ apps/frontend/src/app/models/plan/create-plan.dto.ts
@@ -1,7 +1,7 @@
-import { Model } from '../utils/types';
-import { Dto } from './base.dto';
-import { CreateGathering } from './create-gathering.dto';
-import { GameDto } from './game.dto';
+import { Model } from '../../utils/types';
+import { Dto } from '../base.dto';
+import { GameDto } from '../game/game.dto';
+import { CreateGathering } from '../gathering/create-gathering.dto';
 
 /**
  * Dto for creating plans.
