+++ apps/frontend/src/app/models/user/register-user.dto.ts
@@ -1,19 +1,21 @@
-import { Model } from '../utils/types';
-import { Dto } from './base.dto';
+import { Model } from '../../utils/types';
+import { Dto } from '../base.dto';
 
 /**
- * Dto for users.
+ * Dto for registering a user.
  *
  * @property {string} username - The username of the user
  * @property {string} firstName - The first name of the user
  * @property {string} lastName - The last name of the user
  * @property {string} email - The email address of the user
+ * @property {string} password - The password of the user
  */
-export class UserDto extends Dto {
+export class RegisterUserDto extends Dto {
   public username!: string;
   public firstName!: string;
   public lastName!: string;
   public email!: string;
+  public password!: string;
 }
 
-export type User = Model<UserDto>;
+export type RegisterUser = Model<RegisterUserDto>;
