+++ apps/frontend/src/app/models/user/user-update.dto.ts
@@ -1,4 +1,4 @@
-import { Model } from '../utils/types';
+import { Model } from '../../utils/types';
 import { UserDto } from './user.dto';
 
 /**
