+++ apps/frontend/src/app/services/user.service.ts
@@ -5,9 +5,9 @@ import { Inject, Injectable } from '@angular/core';
 
 import { API_BASE_URL } from '../app.config';
 import { JwtDto } from '../models/jwt.dto';
-import { PasswordUpdate } from '../models/password-update.dto';
-import { UserUpdate } from '../models/user-update.dto';
-import { UserDto } from '../models/user.dto';
+import { PasswordUpdate } from '../models/user/password-update.dto';
+import { UserUpdate } from '../models/user/user-update.dto';
+import { UserDto } from '../models/user/user.dto';
 import { ResponseHandler } from '../utils/response.handler';
 import { Uuid } from '../utils/types';
